benchmark/requirements.txt
transformers>=4.40.0
Pillow>=10.0.0
numpy>=1.24.0
